+++ package.json
@@ -14,6 +14,7 @@
     "@tauri-apps/api": "^1.2.0",
     "next": "^12.2.5",
     "react": "^18.2.0",
+    "react-csv": "^2.2.2",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
