+++ src/components/molecules/FilteredAccountCount/filtered-account-count.component.tsx
@@ -3,11 +3,11 @@ import { AccountMap } from "../../../types";
 export const FilteredAccountCount = ({
   accountMap
 }: {
-  accountMap: AccountMap | undefined;
+  accountMap: AccountMap | null;
 }): JSX.Element => (
   <div>
     <h2>Count of Filtered Accounts</h2>
-    {accountMap !== undefined &&
+    {accountMap !== null &&
       Object.keys(accountMap).length > 0 &&
       Object.keys(accountMap).length}
   </div>
