+++ src/pages/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useHome } from './home.hooks';
