+++ src/types/mode.ts
@@ -1,0 +1,1 @@
+export type PageMode = "import" | "filter" | "export";
